# Improve serialization performance