… core/src/main/java/io/lionweb/lioncore/java/serialization/JsonSerialization.java
… core/src/main/java/io/lionweb/lioncore/java/serialization/NodePopulator.java
… core/src/main/java/io/lionweb/lioncore/java/serialization/SerializationStatus.java
… core/src/test/java/io/lionweb/lioncore/java/PerformanceTest.java
… core/src/test/java/io/lionweb/lioncore/java/serialization/NodePopulatorTest.java
… core/src/test/java/io/lionweb/lioncore/java/serialization/PerformanceTestOnSerialization.java