Issue #206 reports the measurements obtained for the serialization and deserialization of a large file.

* deserialization: 522ms to 545ms
* serialization: 6595ms to 6685ms

With this PR we get:

* deserialization: 217ms to 358ms
* serialization: 123ms to 130ms

Considering the best case, the deserialization is down to 42% the original time and the serialization is down to 2% of the original time (i.e., we got over a 50x improvement).

We do that mainly by ntroducing `SerializationStatus` and adding features to `DeserializationStatus`. These two classes caches expensive operations, like calculating all the features of a classifier.

We also replace some calls to `Objects.requireNonNull` with ifs, to avoid expensive message calculations.